Issue #30 Centralize all run options in a plugin
